TRE-106: feat(stats): implement get task stats repository

+++ core/model/src/main/java/com/doannd3/treetask/core/model/stats/TaskStats.kt
@@ -6,6 +6,6 @@ data class TaskStats(
     val inProgress: Int,
     val pending: Int,
     val done: Int,
-    val completionRate: Int,
+    val completionRate: Double,
     val recentTasks: List<RecentTaskSummary>,
 )
